added header for global class and got jump working
